:construction: move NavBar in AppBar
